backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pyod
python-multipart
python-dotenv
httpx
google-genai
pydantic
pydantic-settings
statsmodels
SQLAlchemy>=2.0.0
asyncpg>=0.29.0
torch==2.7.1
alembic>=1.13.1
slowapi>=0.1.9
gunicorn>=23.0.0
